notebooks/fl-diabetes-prediction/fl-diabetes-prediction/pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "fl-diabetes-prediction"
version = "1.0.0"
requires-python = ">=3.12"
description = "Federated Learning for Diabetes Prediction"
license = "Apache-2.0"
dependencies = [
    "flwr-datasets>=0.5.0",
    "torch>=2.8.0",
    "imblearn",
    "pandas",
    "scikit-learn==1.6.1",
    "loguru",
    "jupyter",
    "freezegun==1.5.5",
    "syft_flwr==0.4.0",
]

[tool.uv.sources.syft-flwr]
path = "/Users/khoaguin/Desktop/projects/OpenMined/syft-flwr"
editable = true

[tool.hatch.build.targets.wheel]
packages = [
    ".",
]

[tool.flwr.app]
publisher = "OpenMined"

[tool.flwr.app.components]
serverapp = "fl_diabetes_prediction.server_app:app"
clientapp = "fl_diabetes_prediction.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 2
partition-id = 0
num-partitions = 1

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation.options]
num-supernodes = 2

[tool.syft_flwr]
app_name = "ds@openmined.org_fl-diabetes-prediction_1760437649"
datasites = [
    "do1@openmined.org",
    "do2@openmined.org",
]
aggregator = "ds@openmined.org"
